+++ skaffold.yaml
@@ -1159,6 +1159,7 @@ build:
             - pkg/nodeagent/apis/config/validation
             - pkg/nodeagent/bootstrap
             - pkg/nodeagent/controller
+            - pkg/nodeagent/controller/lease
             - pkg/nodeagent/controller/node
             - pkg/nodeagent/controller/operatingsystemconfig
             - pkg/nodeagent/controller/token
